[refactor/#206] 온보딩 수입 버튼 공통 컴포넌트화
